[TASK-211] <<fixed>> rename QuestionDetail to CommunityChallengeQuestionDetail
