+++ itp-front/src/components/CommunityChallengeQuestionDetail.vue
