+++ itp-front/src/views/AnswerQuestionPage.vue
@@ -22,7 +22,7 @@
   import Goal5LearningBanner from '../components/Goal5LearningBanner.vue';
   import MainFooter from '../components/MainFooter.vue';
   import SideBar from '../components/SideBar.vue';
-  import QuestionDetail from '../components/QuestionDetail.vue';
+  import QuestionDetail from '../components/CommunityChallengeQuestionDetail.vue';
   import Header from '../components/Header.vue';
   
   export default {
